Move wait_for_chart_canvas to file level for reuse
Move the helper function from st_pydeck_chart_test.py to
shared/pydeck_utils.py for reuse across pydeck tests.
